Index search

See merge request linayipromoter/promoter!67

+++ linayi.service/src/main/java/com/linayi/service/order/impl/OrderServiceImpl.java
@@ -97,7 +97,7 @@ public class OrderServiceImpl implements OrderService {
 
     @Transactional
     @Override
-    public ResponseData addOrder(Map<String, Object> param) throws Exception {
+    public ResponseData addOrder(Map<String, Object> param){
         Integer userId = (Integer) param.get("userId");
         User user = userMapper.selectUserByuserId(userId);
         Integer receiveAddressId = user.getDefaultReceiveAddressId();
@@ -146,7 +146,10 @@ public class OrderServiceImpl implements OrderService {
         //获取所有的购物车
         List<ShoppingCar> shoppingCars = getShoppingCars(userId, receiveAddressId);
         if (shoppingCars == null){
-            throw new Exception();
+            ShoppingCar shoppingCar = new ShoppingCar();
+            shoppingCar.setQuantity(Integer.valueOf(param.get("quantity") + ""));
+            shoppingCar.setGoodsSkuId(Integer.valueOf(param.get("goodsSkuId") + ""));
+            shoppingCars.add(shoppingCar);
         }
 
         //获取收货地址
@@ -192,7 +195,9 @@ public class OrderServiceImpl implements OrderService {
             GoodsSku goodsSku = goodsSkuMapper.getGoodsById(car.getGoodsSkuId());
             goodsSku.setSoldNum(goodsSku.getSoldNum() == null ? 0 : goodsSku.getSoldNum() + car.getQuantity());
             goodsSkuMapper.update(goodsSku);
-            shoppingCarMapper.deleteCarById(Integer.parseInt(car.getShoppingCarId() + ""));
+            if(car.getShoppingCarId() != null){
+                shoppingCarMapper.deleteCarById(Integer.parseInt(car.getShoppingCarId() + ""));
+            }
         }
         return new ResponseData("success");
     }
@@ -733,7 +738,8 @@ public class OrderServiceImpl implements OrderService {
      * @param areaCode
      * @return
      */
-    private String getAreaNameByAreaCode(String areaCode) {
+    @Override
+    public String getAreaNameByAreaCode(String areaCode) {
         //获取街道名
         String streetName = areaMapper.getNameByCode(areaCode);
         //获取省市区区名
